refactor(orders): move the request registry and status rules off the bridge
First step of #120. `rust/src/api/` is the FRB bridge surface — the only
Rust Dart calls directly — and `api/orders.rs` had grown to hold the
protocol internals too. Neither piece moved here is reachable from Dart.

`mostro::pending` — the correlation registry: `DaemonReply`,
`PendingRequestKind`, `PendingRequest`, the `PENDING_REQUESTS` static and
its matchers, plus `classify_take_reply`, `may_reconcile_stored_id` and
`order_content_key`. Self-contained bookkeeping over its own state.

`mostro::status` — the pure mapping and reconciliation rules:
`status_for_action`, `map_core_status`, `is_hard_terminal`,
`is_terminal_status`, `wire_status_applies`,
`cancellation_wipes_history`, `add_invoice_sync`. `classify_take_reply`
depends on two of these, so extracting the registry without them would
have left a back-dependency from `mostro` into `api`.

Behaviour-preserving: functions moved verbatim, only visibility widened
to `pub(crate)`. `api/orders.rs` 5014 → 4515 lines.

Tests follow their subjects. `request_id_only_matches_the_exact_nonce`
moved to `mostro::pending`, and both modules gained the coverage they
could not easily have while buried in a 5000-line file:

  - pending: only our own pending local UUID may be rebound; the content
    key separates range from fixed orders and buys from sells.
  - status: a wire status may not walk a live trade backwards;
    `SettledHoldInvoice` is terminal but not hard-terminal (conflating
    them would drop a trade still awaiting payout); only pre-trade
    cancellations wipe history; the bond statuses map to None rather
    than to something wrong; actions that carry no status change return
    None rather than a guess.

cargo test 261 passed, clippy clean, flutter analyze unchanged (7
pre-existing), flutter test 229 passed.
