+++ rust/src/mostro/mod.rs
@@ -1,9 +1,11 @@
 pub mod actions;
 pub mod escrow_mode;
 pub mod fsm;
+pub mod pending;
 pub mod pow;
 pub mod protocol_version;
 pub mod session;
+pub mod status;
 
 #[cfg(test)]
 mod cashu_wire;
